Edit the config format
