+++ lib/router.js
@@ -1047,9 +1047,6 @@ router.get('/jpmorganchase', require('./routes/jpmorganchase/research'));
 // 美拍
 router.get('/meipai/user/:uid', require('./routes/meipai/user'));
 
-// 通用配置路由
-router.get('/common/:config', require('./routes/selector-common'));
-
 // 多知网
 router.get('/duozhi', require('./routes/duozhi'));
 
@@ -1121,5 +1118,9 @@ router.get('/luogu/daily/:id?', require('./routes/luogu/daily'));
 
 // 决胜网
 router.get('/juesheng', require('./routes/juesheng'));
+
+// 裏垢女子まとめ
+router.get('/uraaka-joshi', require('./routes/uraaka-joshi/uraaka-joshi'));
+router.get('/uraaka-joshi/:id', require('./routes/uraaka-joshi/uraaka-joshi-user'));
 
 module.exports = router;
