+++ lib/routes/uraaka-joshi/uraaka-joshi-user.js
@@ -1,6 +1,9 @@
-module.exports = (params) => {
+const buildData = require('../../utils/common-config');
+
+module.exports = async (ctx) => {
+    const params = ctx.params;
     const link = `https://www.uraaka-joshi.com/users/${params.id}`;
-    return {
+    ctx.state.data = await buildData({
         link,
         url: link,
         title: `$('.top-profile-card-name-link').text() + '@${params.id} - 裏垢女子まとめ'`,
@@ -12,5 +15,5 @@ module.exports = (params) => {
             pubDate: `new Date($('.post-time').attr('datetime')).toUTCString()`,
             guid: `new Date($('.post-time').attr('datetime')).getTime()`,
         },
-    };
+    });
 };
