+++ lib/routes/uraaka-joshi/uraaka-joshi.js
@@ -1,9 +1,14 @@
-module.exports = () => {
+const buildData = require('../../utils/common-config');
+
+module.exports = async (ctx) => {
     const link = `https://www.uraaka-joshi.com/`;
-    return {
+    ctx.state.data = await buildData({
         link,
         url: link,
-        title: `裏垢女子まとめ`,
+        title: `%title%`,
+        params: {
+            title: '裏垢女子まとめ',
+        },
         item: {
             item: '.content-main .stream .stream-item',
             title: `$('.post-account-group').text() + ' - %title%'`,
@@ -12,5 +17,5 @@ module.exports = () => {
             pubDate: `new Date($('.post-time').attr('datetime')).toUTCString()`,
             guid: `new Date($('.post-time').attr('datetime')).getTime()`,
         },
-    };
+    });
 };
